Increased string lengths and capacities by sizeof(uint64), implemented some string operations

+++ HighC/include/DataStructures/Strings.h
@@ -12,8 +12,8 @@ extern "C"
 typedef struct
 {
 	char* data;
-	uint32_t length;
-	uint32_t capacity;
+	uint64_t length;
+	uint64_t capacity;
 }hcString;
 
 hcString hcNewString(const char* initialString);
@@ -21,16 +21,16 @@ void hcAppendCharToString(hcString* string, char character);
 void hcAppendStringToString(hcString* firstString, hcString secondString);
 void hcAssignCharToString(hcString* string, char character);
 void hcAssignStringToString(hcString* stringCopy, hcString* stringOriginal);
-void hcInsertCharIntoString(hcString* stringTarget, char character, uint32_t index);
-void hcInsertStringIntoString(hcString* stringTarget, char subString, uint32_t index);
-void hcEraseFromString(hcString* string, uint32_t first, uint32_t last);
-void hcReplaceToString(hcString* stringTarget, uint32_t first, uint32_t last, hcString* subString);
+void hcInsertCharIntoString(hcString* stringTarget, char character, uint64_t index);
+void hcInsertStringIntoString(hcString* stringTarget, char subString, uint64_t index);
+void hcEraseFromString(hcString* string, uint64_t first, uint64_t last);
+void hcReplaceToString(hcString* stringTarget, uint64_t first, uint64_t last, hcString* subString);
 void hcSwapStrings(hcString* firstString, hcString* secondString);
 char hcPopBackString(hcString* string);
-hcString hcGetSubString(hcString* string, uint32_t index);
+hcString hcGetSubString(hcString* string, uint64_t index);
 bool hcIsStringAndCharArrayTheSame(hcString* string, const char* text);
 bool hcisStringEmpty(hcString* string);
-void hcReserveString(hcString* string, uint32_t newCapacity);
+void hcReserveString(hcString* string, uint64_t newCapacity);
 void hcShrinkStringToFit(hcString* string);
 void hcClearString(hcString* string);
 void hcFreeString(hcString* string);
